bgprecorder/pyproject.toml
[tool.poetry]
name = "bgprecorder"
version = "0.1.0"
description = ""
authors = ["yas-nyan <yas-nyan@sfc.wide.ad.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
logzero = "^1.7.0"
psycopg2-binary = "^2.9.3"
pickleDB = "^0.9.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
